some package cleanup for service/m365
